Basic MetaTables work.
