+++ MoinMoin/metadata/wikitextutil.py
@@ -5,6 +5,7 @@ from MoinMoin import config
 
 from MoinMoin.Page import LinkCollectingPage
 from MoinMoin.parser.link_collect import Parser as lcparser
+from MoinMoin.parser.text_moin_wiki import Parser
 from MoinMoin.formatter.nullformatter import Formatter as nullformatter
 from MoinMoin.wikiutil import get_processing_instructions, AbsPageName
 
@@ -762,3 +763,34 @@ def replace_metas(request, text, oldmeta, newmeta):
     # Add enter to the end of the line, as it was removed in the
     # beginning of this function, not doing so causes extra edits.
     return text.rstrip() + '\n'
+
+def format_wikitext(request, data, parser=None):
+    request.page.formatter = request.formatter
+    request.formatter.page = request.page
+
+    if not parser:
+        parser = Parser(data, request)
+    else:
+        parser.raw = data
+    parser.request = request
+
+    # Do not store pagelinks for values in metadata listings
+    plstore = getattr(request.formatter, '_store_pagelinks', 0)
+    request.formatter._store_pagelinks = 0
+
+    parser.formatter = request.formatter
+    # No line anchors of any type to table cells
+    request.page.formatter.in_p = 1
+    parser._line_anchordef = lambda: ''
+
+    # Do not parse macros from revision pages. For some reason,
+    # it spawns multiple requests, which are not finished properly,
+    # thus littering a number of readlocks. Besides, the macros do not
+    # return anything useful anyway for pages they don't recognize
+    if '?action=recall' in request.page.page_name:
+        parser._macro_repl = lambda x: x
+
+    out = parser.scan(data, inhibit_p=True)
+
+    request.formatter._store_pagelinks = plstore
+    return out.strip()
